Renames class library to "AnonymousTokens"
