+++ samples/ClientServer/Client/Client.Console/Client.Console.csproj
@@ -11,7 +11,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\..\..\src\AnonymousTokens.Client\AnonymousTokens.Client.csproj" />
-    <ProjectReference Include="..\..\..\..\src\AnonymousTokens.Shared\AnonymousTokens.Shared.csproj" />
+    <ProjectReference Include="..\..\..\..\src\AnonymousTokens\AnonymousTokens.csproj" />
   </ItemGroup>
 
 </Project>
