+++ samples/ClientServer/Server/Server.TokenVerification.Api/Controllers/TokenController.cs
@@ -1,7 +1,7 @@
 
 using AnonymousTokens.Server.Protocol;
-using AnonymousTokens.Shared.Services;
-using AnonymousTokens.Shared.Services.InMemory;
+using AnonymousTokens.Services;
+using AnonymousTokens.Services.InMemory;
 
 using Microsoft.AspNetCore.Mvc;
