+++ src/AnonymousTokens.Client/AnonymousTokens.Client.csproj
@@ -16,6 +16,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AnonymousTokens.Shared\AnonymousTokens.Shared.csproj" />
+    <ProjectReference Include="..\AnonymousTokens\AnonymousTokens.csproj" />
   </ItemGroup>
 </Project>
