+++ src/AnonymousTokens.Client/Protocol/Initiator.cs
@@ -1,4 +1,3 @@
-using AnonymousTokens.Shared;
 
 using Org.BouncyCastle.Asn1.X9;
 using Org.BouncyCastle.Crypto.Parameters;
