--- src/AnonymousTokens.Shared/Services/InMemory/EmbeddedResourceConstants.cs
@@ -1,7 +1,0 @@
-namespace AnonymousTokens.Shared.Services.InMemory
-{
-    internal static class EmbeddedResourceConstants
-    {
-        internal const string ResourceBasePath = "AnonymousTokens.Shared.Services.InMemory.";
-    }
-}
