+++ src/AnonymousTokens/CPChallengeGenerator.cs
@@ -8,7 +8,7 @@ using System.Text;
 
 using ECPoint = Org.BouncyCastle.Math.EC.ECPoint;
 
-namespace AnonymousTokens.Shared
+namespace AnonymousTokens
 {
     public static class CPChallengeGenerator
     {
