+++ src/AnonymousTokens/Services/InMemory/EmbeddedResourceConstants.cs
@@ -1,0 +1,7 @@
+namespace AnonymousTokens.Services.InMemory
+{
+    internal static class EmbeddedResourceConstants
+    {
+        internal const string ResourceBasePath = "AnonymousTokens.Services.InMemory.";
+    }
+}
