+++ src/AnonymousTokens/Services/InMemory/InMemoryPrivateKeyStore.cs
@@ -2,7 +2,7 @@ using Org.BouncyCastle.Crypto;
 using Org.BouncyCastle.Crypto.Parameters;
 using Org.BouncyCastle.Math;
 
-namespace AnonymousTokens.Shared.Services.InMemory
+namespace AnonymousTokens.Services.InMemory
 {
     public class InMemoryPrivateKeyStore : IPrivateKeyStore
     {
