+++ src/AnonymousTokens/Services/InMemory/InMemorySeedStore.cs
@@ -3,7 +3,7 @@ using Org.BouncyCastle.Utilities.Encoders;
 
 using System.Collections.Generic;
 
-namespace AnonymousTokens.Shared.Services.InMemory
+namespace AnonymousTokens.Services.InMemory
 {
     public class InMemorySeedStore : ISeedStore
     {
